requirements.txt
python-dotenv 
langchain
google-generativeai
langchain-google-genai
langchain-community
pillow
pypdf2
chroma  
langchainhub
langchain-chroma

-e .